Improve partial inference of key container operations (#37084)
`resize!` and `sizehint!` are big targets for invalidation, and quite
a lot of code was calling these methods with poor inference about the
size. Since the size is just `Int`, asserting that eliminates numerous
vulnerabilities. Likewise, bounds-checking always works with `Bool`.
Finally, add a couple of short-circuits for `eltype(::Tuple{T,...})`,
including one that bypasses `typejoin`.

+++ base/array.jl
@@ -984,7 +984,7 @@ push!(a::AbstractVector, iter...) = append!(a, iter)
 function _append!(a, ::Union{HasLength,HasShape}, iter)
     n = length(a)
     i = lastindex(a)
-    resize!(a, n+length(iter))
+    resize!(a, n+Int(length(iter))::Int)
     @inbounds for (i, item) in zip(i+1:lastindex(a), iter)
         a[i] = item
     end
